requirements.txt
alembic==0.9.9
amqp==2.3.2
aniso8601==3.0.0
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.3
blinker==1.4
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
click==6.7
cssmin==0.2.0
defusedxml==0.5.0
dominate==2.3.1
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Assets==0.12
Flask-AutoIndex==0.6
Flask-Babel==0.11.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Celery-Helper==1.1.0
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
flask-nav==0.6
Flask-OpenID==1.2.5
Flask-Principal==0.4.0
Flask-REST-JSONAPI==0.19.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Silk==0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
itsdangerous==0.24
Jinja2==2.10
jsmin==2.2.2
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.1
marshmallow-jsonapi==0.19.0.post0
py-postgresql==1.2.1
pycparser==2.18
PyMySQL==0.8.1
python-dateutil==2.7.3
python-editor==1.0.3
python3-openid==3.1.0
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.7
uuid==1.30
vine==1.1.4
visitor==0.1.3
webassets==0.12.1
Werkzeug==0.14.1
WTForms==2.1
